hardware/requirements.txt
pyserial>=3.5
mediapipe>=0.10.13
opencv-python>=4.9.0.80
numpy>=1.24.3

